conditionals for products

+++ test_products.sh
@@ -9,27 +9,45 @@ NEW_PRODUCT=$(curl -s -X POST -H "Content-Type: application/json" -d '{"product_
 NEW_PRODUCT_ID=$(echo $NEW_PRODUCT | jq '.id')
 echo -e "ID: $NEW_PRODUCT_ID Name: Test Product"
 
-# get '/:id' : Get newly made product by ID
-echo -e "\033[1;4mTesting GET a specific product by ID:\033[0m"
-curl -s -X GET $BASE_URL/$NEW_PRODUCT_ID | jq
+if [ $NEW_PRODUCT_ID ]; then 
 
-# put /:id' : Update the product name to Updated Test Product
-echo -e "\033[1;4mTesting PUT (update) a product by ID:\033[0m"
-HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" -X PUT -H "Content-Type: application/json" -d '{"product_name": "Updated Test Product"}' $BASE_URL/$NEW_PRODUCT_ID)
-echo "HTTP Status: $HTTP_STATUS"
+    # get '/:id' : Get newly made product by ID
+    echo -e "\033[1;4mTesting GET a specific product by ID:\033[0m"
+    curl -s -X GET $BASE_URL/$NEW_PRODUCT_ID | jq
 
-# get '/:id' : Get updated product by ID
-echo -e "\033[1;4mTesting GET the updated product:\033[0m"
-curl -s -X GET $BASE_URL/$NEW_PRODUCT_ID | jq
+    # put /:id' : Update the product name to Updated Test Product
+    echo -e "\033[1;4mTesting PUT (update) a product by ID:\033[0m"
+    HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" -X PUT -H "Content-Type: application/json" -d '{"product_name": "Updated Test Product"}' $BASE_URL/$NEW_PRODUCT_ID)
+    echo "HTTP Status: $HTTP_STATUS"
 
-# delete '/:id' : Delete product by ID
-echo -e "\033[1;4mTesting DELETE a product by ID:\033[0m"
-HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" -X DELETE $BASE_URL/$NEW_PRODUCT_ID)
-echo "HTTP Status: $HTTP_STATUS"
+    if [ $HTTP_STATUS -eq 200 ] || [ $HTTP_STATUS -eq 204 ]; then
 
-# get '/:id' : expect an error : Get deleted product by ID 
-echo -e "\033[1;4mTesting GET the deleted product (expecting an error):\033[0m"
-curl -s -X GET $BASE_URL/$NEW_PRODUCT_ID | jq
+        # get '/:id' : Get updated product by ID
+        echo -e "\033[1;4mTesting GET the updated product:\033[0m"
+        curl -s -X GET $BASE_URL/$NEW_PRODUCT_ID | jq
+
+        # delete '/:id' : Delete product by ID
+        echo -e "\033[1;4mTesting DELETE a product by ID:\033[0m"
+        HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" -X DELETE $BASE_URL/$NEW_PRODUCT_ID)
+        echo "HTTP Status: $HTTP_STATUS"
+
+        if [ $HTTP_STATUS -eq 200 ] || [ $HTTP_STATUS -eq 204 ]; then
+
+            # get '/:id' : expect an error : Get deleted product by ID 
+            echo -e "\033[1;4mTesting GET the deleted product (expecting an error):\033[0m"
+            curl -s -X GET $BASE_URL/$NEW_PRODUCT_ID | jq
+
+        else
+            echo "Delete request caused an error"
+        fi
+
+    else
+        echo "Put request caused an error"
+    fi
+
+else
+  echo "Post request caused an error"
+fi
 
 # get '/' : GetAll : Verify deletion against all products 
 echo -e "\033[1;4mTesting GET all products:\033[0m"
